[#57] Smoke CRUD domain and data shared  - Add, Edit and Delete smoke domain and data members moved into a new library to be reused by smoke list history.
