--- features/home/data/src/main/java/com/feragusper/smokeanalytics/features/home/data/FirestoreModule.kt
@@ -1,16 +1,0 @@
-package com.feragusper.smokeanalytics.features.home.data
-
-import com.google.firebase.firestore.FirebaseFirestore
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@InstallIn(SingletonComponent::class)
-@Module
-internal object FirestoreModule {
-    @Singleton
-    @Provides
-    fun provideFirestore(): FirebaseFirestore = FirebaseFirestore.getInstance()
-}
