--- features/home/data/src/main/java/com/feragusper/smokeanalytics/features/home/data/SmokeEntity.kt
@@ -1,5 +1,0 @@
-package com.feragusper.smokeanalytics.features.home.data
-
-import java.util.Date
-
-data class SmokeEntity(val date: Date)
