--- features/home/domain/src/main/java/com/feragusper/smokeanalytics/features/home/domain/AddSmokeUseCase.kt
@@ -1,7 +1,0 @@
-package com.feragusper.smokeanalytics.features.home.domain
-
-import javax.inject.Inject
-
-class AddSmokeUseCase @Inject constructor(private val smokeRepository: SmokeRepository) {
-    suspend operator fun invoke() = smokeRepository.addSmoke()
-}
