--- features/home/domain/src/main/java/com/feragusper/smokeanalytics/features/home/domain/DeleteSmokeUseCase.kt
@@ -1,8 +1,0 @@
-package com.feragusper.smokeanalytics.features.home.domain
-
-import java.util.Date
-import javax.inject.Inject
-
-class DeleteSmokeUseCase @Inject constructor(private val smokeRepository: SmokeRepository) {
-    suspend operator fun invoke(id: String) = smokeRepository.deleteSmoke(id)
-}
