--- features/home/domain/src/main/java/com/feragusper/smokeanalytics/features/home/domain/Smoke.kt
@@ -1,9 +1,0 @@
-package com.feragusper.smokeanalytics.features.home.domain
-
-import java.util.Date
-
-data class Smoke(
-    val id: String,
-    val date: Date,
-    val timeElapsedSincePreviousSmoke: Pair<Long, Long>
-)
