--- features/home/domain/src/main/java/com/feragusper/smokeanalytics/features/home/domain/SmokeRepository.kt
@@ -1,10 +1,0 @@
-package com.feragusper.smokeanalytics.features.home.domain
-
-import java.util.Date
-
-interface SmokeRepository {
-    suspend fun addSmoke()
-    suspend fun fetchSmokes(): List<Smoke>
-    suspend fun editSmoke(id: String, date: Date)
-    suspend fun deleteSmoke(id: String)
-}
