+++ features/home/presentation/src/main/java/com/feragusper/smokeanalytics/features/home/presentation/process/HomeProcessHolder.kt
@@ -1,8 +1,5 @@
 package com.feragusper.smokeanalytics.features.home.presentation.process
 
-import com.feragusper.smokeanalytics.features.home.domain.AddSmokeUseCase
-import com.feragusper.smokeanalytics.features.home.domain.DeleteSmokeUseCase
-import com.feragusper.smokeanalytics.features.home.domain.EditSmokeUseCase
 import com.feragusper.smokeanalytics.features.home.domain.FetchSmokeCountListUseCase
 import com.feragusper.smokeanalytics.features.home.presentation.mvi.HomeIntent
 import com.feragusper.smokeanalytics.features.home.presentation.mvi.HomeResult
@@ -11,6 +8,9 @@ import com.feragusper.smokeanalytics.libraries.architecture.presentation.extensi
 import com.feragusper.smokeanalytics.libraries.architecture.presentation.process.MVIProcessHolder
 import com.feragusper.smokeanalytics.libraries.authentication.domain.FetchSessionUseCase
 import com.feragusper.smokeanalytics.libraries.authentication.domain.Session
+import com.feragusper.smokeanalytics.libraries.smokes.domain.AddSmokeUseCase
+import com.feragusper.smokeanalytics.libraries.smokes.domain.DeleteSmokeUseCase
+import com.feragusper.smokeanalytics.libraries.smokes.domain.EditSmokeUseCase
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
