+++ features/home/presentation/src/test/java/com/feragusper/smokeanalytics/features/home/presentation/HomeViewModelTest.kt
@@ -1,12 +1,12 @@
 package com.feragusper.smokeanalytics.features.home.presentation
 
 
-import com.feragusper.smokeanalytics.features.home.domain.Smoke
 import com.feragusper.smokeanalytics.features.home.domain.SmokeCountListResult
 import com.feragusper.smokeanalytics.features.home.presentation.mvi.HomeIntent
 import com.feragusper.smokeanalytics.features.home.presentation.mvi.HomeResult
 import com.feragusper.smokeanalytics.features.home.presentation.mvi.HomeViewState
 import com.feragusper.smokeanalytics.features.home.presentation.process.HomeProcessHolder
+import com.feragusper.smokeanalytics.libraries.smokes.domain.Smoke
 import io.mockk.every
 import io.mockk.mockk
 import kotlinx.coroutines.Dispatchers
