+++ libraries/smokes/data/src/main/java/com/feragusper/smokeanalytics/libraries/smokes/data/SmokeEntity.kt
@@ -1,0 +1,5 @@
+package com.feragusper.smokeanalytics.libraries.smokes.data
+
+import java.util.Date
+
+data class SmokeEntity(val date: Date)
