+++ libraries/smokes/data/src/main/java/com/feragusper/smokeanalytics/libraries/smokes/data/SmokeRepositoryImpl.kt
@@ -1,0 +1,70 @@
+package com.feragusper.smokeanalytics.libraries.smokes.data
+
+import com.feragusper.smokeanalytics.libraries.architecture.domain.helper.timeAfter
+import com.feragusper.smokeanalytics.libraries.smokes.data.SmokeRepositoryImpl.FirestoreCollection.Companion.SMOKES
+import com.feragusper.smokeanalytics.libraries.smokes.data.SmokeRepositoryImpl.FirestoreCollection.Companion.USERS
+import com.feragusper.smokeanalytics.libraries.smokes.domain.Smoke
+import com.feragusper.smokeanalytics.libraries.smokes.domain.SmokeRepository
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.firestore.DocumentSnapshot
+import com.google.firebase.firestore.FirebaseFirestore
+import com.google.firebase.firestore.Query.Direction
+import kotlinx.coroutines.tasks.await
+import java.util.Date
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class SmokeRepositoryImpl @Inject constructor(
+    private val firebaseFirestore: FirebaseFirestore,
+    private val firebaseAuth: FirebaseAuth,
+) : SmokeRepository {
+
+    interface FirestoreCollection {
+        companion object {
+            const val USERS = "users"
+            const val SMOKES = "smokes"
+        }
+    }
+
+    override suspend fun addSmoke() {
+        smokes().add(SmokeEntity(Date())).await()
+    }
+
+    private fun smokes() = firebaseAuth.currentUser?.uid?.let {
+        firebaseFirestore.collection("$USERS/$it/$SMOKES")
+    } ?: throw IllegalStateException("User not logged in")
+
+    override suspend fun editSmoke(id: String, date: Date) {
+        smokes()
+            .document(id)
+            .set(SmokeEntity(date))
+            .await()
+    }
+
+    override suspend fun deleteSmoke(id: String) {
+        smokes()
+            .document(id)
+            .delete()
+            .await()
+    }
+
+    override suspend fun fetchSmokes(): List<Smoke> {
+        val smokes = smokes()
+            .orderBy(Smoke::date.name, Direction.DESCENDING)
+            .get()
+            .await()
+
+        return smokes.documents.mapIndexedNotNull { index, document ->
+            Smoke(
+                id = document.id,
+                date = document.getDate(),
+                timeElapsedSincePreviousSmoke = document.getDate()
+                    .timeAfter(smokes.documents.getOrNull(index + 1)?.getDate()),
+            )
+        }
+    }
+
+    private fun DocumentSnapshot.getDate() =
+        getDate(Smoke::date.name) ?: throw IllegalStateException("Date not found")
+}
