+++ libraries/smokes/data/src/main/java/com/feragusper/smokeanalytics/libraries/smokes/data/SmokeRepositoryModule.kt
@@ -1,0 +1,14 @@
+package com.feragusper.smokeanalytics.libraries.smokes.data
+
+import com.feragusper.smokeanalytics.libraries.smokes.domain.SmokeRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ViewModelComponent
+
+@Module
+@InstallIn(ViewModelComponent::class)
+abstract class SmokeRepositoryModule {
+    @Binds
+    abstract fun provideSmokeRepository(smokeRepository: SmokeRepositoryImpl): SmokeRepository
+}
