+++ libraries/smokes/data/src/main/java/com/feragusper/smokeanalytics/libraries/smokes/data/di/FirestoreModule.kt
@@ -1,0 +1,16 @@
+package com.feragusper.smokeanalytics.libraries.smokes.data.di
+
+import com.google.firebase.firestore.FirebaseFirestore
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@InstallIn(SingletonComponent::class)
+@Module
+internal object FirestoreModule {
+    @Singleton
+    @Provides
+    fun provideFirestore(): FirebaseFirestore = FirebaseFirestore.getInstance()
+}
