+++ libraries/smokes/domain/build.gradle.kts
@@ -5,7 +5,6 @@ plugins {
 
 dependencies {
     implementation(project(":libraries:architecture:domain"))
-    implementation(project(":libraries:smokes:domain"))
     implementation(libs.javax.inject)
     testImplementation(platform(libs.junit.bom))
     testImplementation(libs.bundles.test)
