+++ libraries/smokes/domain/src/main/java/com/feragusper/smokeanalytics/libraries/smokes/domain/AddSmokeUseCase.kt
@@ -1,0 +1,7 @@
+package com.feragusper.smokeanalytics.libraries.smokes.domain
+
+import javax.inject.Inject
+
+class AddSmokeUseCase @Inject constructor(private val smokeRepository: SmokeRepository) {
+    suspend operator fun invoke() = smokeRepository.addSmoke()
+}
