+++ libraries/smokes/domain/src/main/java/com/feragusper/smokeanalytics/libraries/smokes/domain/DeleteSmokeUseCase.kt
@@ -1,0 +1,7 @@
+package com.feragusper.smokeanalytics.libraries.smokes.domain
+
+import javax.inject.Inject
+
+class DeleteSmokeUseCase @Inject constructor(private val smokeRepository: SmokeRepository) {
+    suspend operator fun invoke(id: String) = smokeRepository.deleteSmoke(id)
+}
