+++ libraries/smokes/domain/src/main/java/com/feragusper/smokeanalytics/libraries/smokes/domain/Smoke.kt
@@ -1,0 +1,9 @@
+package com.feragusper.smokeanalytics.libraries.smokes.domain
+
+import java.util.Date
+
+data class Smoke(
+    val id: String,
+    val date: Date,
+    val timeElapsedSincePreviousSmoke: Pair<Long, Long>
+)
