+++ libraries/smokes/domain/src/main/java/com/feragusper/smokeanalytics/libraries/smokes/domain/SmokeRepository.kt
@@ -1,0 +1,10 @@
+package com.feragusper.smokeanalytics.libraries.smokes.domain
+
+import java.util.Date
+
+interface SmokeRepository {
+    suspend fun addSmoke()
+    suspend fun fetchSmokes(): List<Smoke>
+    suspend fun editSmoke(id: String, date: Date)
+    suspend fun deleteSmoke(id: String)
+}
