+++ settings.gradle.kts
@@ -17,17 +17,18 @@ dependencyResolutionManagement {
 
 rootProject.name = "SmokeAnalytics"
 include(":app")
-include(":libraries:design")
-include(":libraries:architecture:presentation")
-include(":libraries:architecture:domain")
-include(":libraries:authentication:presentation")
-include(":libraries:authentication:domain")
-include(":libraries:authentication:data")
-include(":features:settings:presentation")
-include(":features:home:presentation")
-include(":features:home:data")
-include(":features:home:domain")
-include(":features:stats:presentation")
 include(":features:devtools:presentation")
+include(":features:home:domain")
+include(":features:home:presentation")
+include(":features:settings:presentation")
+include(":features:stats:presentation")
+include(":libraries:architecture:domain")
+include(":libraries:architecture:presentation")
+include(":libraries:authentication:data")
+include(":libraries:authentication:domain")
+include(":libraries:authentication:presentation")
+include(":libraries:design")
+include(":libraries:smokes:data")
+include(":libraries:smokes:domain")
